test(sync): keep estimate coverage in protected inventory
